+++ pcs/lib/booth/constants.py
@@ -1,6 +1,10 @@
-DEFAULT_INSTANCE_NAME = "booth"
+from pcs import settings
+from pcs.common.types import StringCollection
 
-GLOBAL_KEYS = (
+DEFAULT_INSTANCE_NAME = "booth"
+AUTHFILE_FIX_OPTION = "enable-authfile"
+
+GLOBAL_KEYS: StringCollection = (
     "transport",
     "port",
     "name",
@@ -14,6 +18,11 @@ GLOBAL_KEYS = (
     "arbitrator-group",
     "debug",
     "ticket",
+) + (
+    (AUTHFILE_FIX_OPTION,)
+    if settings.booth_enable_authfile_set_enabled
+    or settings.booth_enable_authfile_unset_enabled
+    else tuple()
 )
 
 TICKET_KEYS = (