+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   in tech-preview state and thus may be changed in the future
 - Support for resource/stonith agent self-validation of instance attributes via
   pacemaker ([rhbz#2112270])
+- Support for booth 'enable-authfile' fix ([rhbz#2116295])
 
 ### Fixed
 - `pcs resource manage --monitor` no longer enables monitor operation for all
@@ -28,6 +29,7 @@
 [rhbz#2092950]: https://bugzilla.redhat.com/show_bug.cgi?id=2092950
 [rhbz#2102663]: https://bugzilla.redhat.com/show_bug.cgi?id=2102663
 [rhbz#2112270]: https://bugzilla.redhat.com/show_bug.cgi?id=2112270
+[rhbz#2116295]: https://bugzilla.redhat.com/show_bug.cgi?id=2116295
 [rhbz#2116841]: https://bugzilla.redhat.com/show_bug.cgi?id=2116841